projects/project-data-pipeline/requirements.txt
pandas
pyyaml
simple-salesforce
python-dotenv
tableauserverclient==0.30.0
tabulate
